# artifacts/api-server/src/routes/analytics.ts
import { Router, type IRouter } from "express";
import { db } from "@workspace/db";
import { workflowsTable, stagesTable, alertsTable } from "@workspace/db/schema";

const router: IRouter = Router();

router.get("/analytics/trends", async (req, res) => {
  try {
    const days = req.query.days ? parseInt(req.query.days as string) : 30;
    const workflows = await db.select().from(workflowsTable);

    const dates: string[] = [];
    const operationalHealth: number[] = [];
    const flowScores: number[] = [];
    const riskScores: number[] = [];
    const improvementScores: number[] = [];
    const executionScores: number[] = [];

    const avgFlow = workflows.length > 0
      ? workflows.reduce((s, w) => s + w.flowScore, 0) / workflows.length
      : 80;
    const avgRisk = workflows.length > 0
      ? workflows.reduce((s, w) => s + w.riskScore, 0) / workflows.length
      : 80;
    const avgImprove = workflows.length > 0
      ? workflows.reduce((s, w) => s + w.improvementScore, 0) / workflows.length
      : 80;
    const avgExec = workflows.length > 0
      ? workflows.reduce((s, w) => s + w.executionScore, 0) / workflows.length
      : 80;

    for (let i = days - 1; i >= 0; i--) {
      const date = new Date();
      date.setDate(date.getDate() - i);
      dates.push(date.toISOString().split("T")[0]);

      const noiseFactor = 0.85 + (i / days) * 0.1;
      const jitter = () => (Math.random() - 0.5) * 8;

      const f = Math.min(100, Math.max(10, Math.round(avgFlow * noiseFactor + jitter())));
      const r = Math.min(100, Math.max(10, Math.round(avgRisk * noiseFactor + jitter())));
      const im = Math.min(100, Math.max(10, Math.round(avgImprove * noiseFactor + jitter())));
      const ex = Math.min(100, Math.max(10, Math.round(avgExec * noiseFactor + jitter())));

      flowScores.push(f);
      riskScores.push(r);
      improvementScores.push(im);
      executionScores.push(ex);
      operationalHealth.push(Math.round((f + r + im + ex) / 4));
    }

    res.json({
      dates,
      operationalHealth,
      flowScore: flowScores,
      riskScore: riskScores,
      improvementScore: improvementScores,
      executionScore: executionScores,
    });
  } catch (err) {
    req.log.error({ err }, "Failed to get trends");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/analytics/workflow-performance", async (req, res) => {
  try {
    const workflows = await db.select().from(workflowsTable);
    const stages = await db.select().from(stagesTable);
    const alerts = await db.select().from(alertsTable);

    const result = workflows.map((wf) => {
      const wfStages = stages.filter((s) => s.workflowId === wf.id);
      const completedStages = wfStages.filter((s) => s.status === "completed");
      const completionRate = wfStages.length > 0
        ? Math.round((completedStages.length / wfStages.length) * 100)
        : 0;

      const stageDurations = wfStages
        .filter((s) => s.startedAt && s.completedAt)
        .map((s) =>
          (new Date(s.completedAt!).getTime() - new Date(s.startedAt!).getTime()) / 86400000
        );
      const avgStageDurationDays =
        stageDurations.length > 0
          ? Math.round(stageDurations.reduce((a, b) => a + b, 0) / stageDurations.length * 10) / 10
          : 0;

      return {
        workflowId: wf.id,
        title: wf.title,
        healthScore: wf.healthScore,
        stoplight: wf.stoplight,
        completionRate,
        avgStageDurationDays,
        bottleneckCount: wfStages.filter((s) => s.isBottleneck).length,
        alertCount: alerts.filter((a) => a.workflowId === wf.id).length,
      };
    });

    res.json(result);
  } catch (err) {
    req.log.error({ err }, "Failed to get workflow performance");
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;