# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ascent 1.0 - Operational Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all assets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, maintenance, retired, critical]
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    put:
      operationId: updateAsset
      tags: [assets]
      summary: Update asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetBody"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: workflowId
          in: query
          schema:
            type: integer
        - name: assetId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create/link a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, warning, info]
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/read:
    patch:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark an alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /analytics/trends:
    get:
      operationId: getAnalyticsTrends
      tags: [analytics]
      summary: Get score trends over time
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendData"
...
components:
  schemas:
...
    Stoplight:
      type: string
      enum: [red, yellow, green]
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        model:
          type: string
          nullable: true
        serial:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, maintenance, retired, critical]
        installDate:
          type: string
          nullable: true
        warrantyStart:
          type: string
          nullable: true
        warrantyExpiration:
          type: string
          nullable: true
        warrantyDaysRemaining:
          type: integer
          nullable: true
        lifeExpectancyYears:
          type: number
          nullable: true
        maintenanceSchedule:
          type: string
          nullable: true
        healthScore:
          type: number
        stoplight:
          $ref: "#/components/schemas/Stoplight"
        location:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - status
        - healthScore
        - stoplight
        - createdAt
...
    CreateAssetBody:
      type: object
      properties:
        name:
          type: string
        model:
          type: string
          nullable: true
        serial:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, maintenance, retired, critical]
        installDate:
          type: string
          nullable: true
        warrantyStart:
          type: string
          nullable: true
        warrantyExpiration:
          type: string
          nullable: true
        lifeExpectancyYears:
          type: number
          nullable: true
        maintenanceSchedule:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
      required: [name]
    UpdateAssetBody:
      type: object
      properties:
        name:
          type: string
        model:
          type: string
          nullable: true
        serial:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, maintenance, retired, critical]
        installDate:
          type: string
          nullable: true
        warrantyStart:
          type: string
          nullable: true
        warrantyExpiration:
          type: string
          nullable: true
        lifeExpectancyYears:
          type: number
          nullable: true
        maintenanceSchedule:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
    Document:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        url:
          type: string
          nullable: true
        detectedType:
          type: string
          nullable: true
        workflowId:
          type: integer
          nullable: true
        stageId:
          type: integer
          nullable: true
        assetId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, filename, createdAt]
    CreateDocumentBody:
      type: object
      properties:
        filename:
          type: string
        url:
          type: string
          nullable: true
        detectedType:
          type: string
          nullable: true
        workflowId:
          type: integer
          nullable: true
        stageId:
          type: integer
          nullable: true
        assetId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [filename]
    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [status_change, due_date, bottleneck, warranty_expiration, risk_escalation, overdue]
        severity:
          type: string
          enum: [critical, warning, info]
        title:
          type: string
        message:
          type: string
        workflowId:
          type: integer
          nullable: true
        assetId:
          type: integer
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - type
        - severity
        - title
        - message
        - isRead
        - createdAt
...
    TrendData:
      type: object
      properties:
        dates:
          type: array
          items:
            type: string
        operationalHealth:
          type: array
          items:
            type: number
        flowScore:
          type: array
          items:
            type: number
        riskScore:
          type: array
          items:
            type: number
        improvementScore:
          type: array
          items:
            type: number
        executionScore:
          type: array
          items:
            type: number
      required:
        - dates
        - operationalHealth
        - flowScore
        - riskScore
        - improvementScore
        - executionScore
...