# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ascent 1.0 - Operational Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { AssetStatus } from "./assetStatus";
import type { Stoplight } from "./stoplight";

export interface Asset {
  id: number;
  name: string;
  model?: string | null;
  serial?: string | null;
  status: AssetStatus;
  installDate?: string | null;
  warrantyStart?: string | null;
  warrantyExpiration?: string | null;
  warrantyDaysRemaining?: number | null;
  lifeExpectancyYears?: number | null;
  maintenanceSchedule?: string | null;
  healthScore: number;
  stoplight: Stoplight;
  location?: string | null;
  createdAt: string;
}