# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ascent 1.0 - Operational Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  filename: string;
  url?: string | null;
  detectedType?: string | null;
  workflowId?: number | null;
  stageId?: number | null;
  assetId?: number | null;
  notes?: string | null;
  createdAt: string;
}